+++ R/model-taxa.R
@@ -159,7 +159,10 @@ getLWCoeffs <- function(taxa_list = NULL, asfis_list = NULL) {
   # 1. Load both databases
   taxa_data <- list(
     fishbase = rfishbase::load_taxa(server = "fishbase"),
-    sealifebase = rfishbase::load_taxa(server = "sealifebase")
+    sealifebase = rfishbase::load_taxa(
+      server = "sealifebase",
+      version = "24.07"
+    )
   )
 
   # 2. Process species list
@@ -330,7 +333,10 @@ get_fao_groups <- function(fao_codes = NULL, asfis_list = NULL) {
 load_taxa_databases <- function() {
   list(
     fishbase = rfishbase::load_taxa(server = "fishbase"),
-    sealifebase = rfishbase::load_taxa(server = "sealifebase")
+    sealifebase = rfishbase::load_taxa(
+      server = "sealifebase",
+      version = "24.07"
+    )
   )
 }
 
@@ -485,7 +491,8 @@ get_species_areas_batch <- function(matched_species) {
     rfishbase::faoareas(
       sealifebase_species,
       fields = "AreaCode",
-      server = "sealifebase"
+      server = "sealifebase",
+      version = "24.07"
     ) %>%
       dplyr::mutate(database = "sealifebase")
   }
@@ -583,7 +590,8 @@ get_length_weight_batch <- function(
     rfishbase::length_weight(
       sealifebase_species,
       fields = c("Species", "SpecCode", "Type", "EsQ", "a", "b"),
-      server = "sealifebase"
+      server = "sealifebase",
+      version = "24.07"
     ) %>%
       dplyr::mutate(database = "sealifebase")
   }
@@ -624,7 +632,8 @@ get_length_weight_batch <- function(
       rfishbase::species(
         sealifebase_species,
         fields = c("Species", "SpecCode", "Length", "CommonLength", "Weight"),
-        server = "sealifebase"
+        server = "sealifebase",
+        version = "24.07"
       ) %>%
         dplyr::mutate(database = "sealifebase")
     }